Tweak subsidy curve tangents
I believe you want tangents only on the starting and ending keys here, otherwise you get weird plateaus every year. Unless that's what you intended, in which case it should probably be a stepwise curve where the value is constant throughout the year?
Also fix using key=12 (which is 1963, not 62).